User Tester Updated

+++ handler/staff.py
@@ -19,37 +19,50 @@ class StaffHandler:
             result_list.append(result)
         return result_list
 
-    def getStaffByEid(self, eid):
+    def getStaffByEid(self, form):
+        eid = v().toLower(form['eid'])
         if eid == '':
-            return False, 'invalid_eid'
+            return False, None, 'invalid_eid'
         dao = StaffDao()
         staff = dao.getStaffByEid(eid)
         if staff == None:
-            return False, 'no_staff'
+            return False, None, 'no_staff'
         else:
-            return staff, 'staff_exists'
+            return True, staff, 'staff_exists'
+
+    def getStaffByEidMain(self, eid):
+        if eid == '':
+            return False, None, 'invalid_eid'
+        dao = StaffDao()
+        staff = dao.getStaffByEid(eid)
+        if staff == None:
+            return False, None, 'no_staff'
+        else:
+            return True, staff, 'staff_exists'
 
     def getStaffByEmail(self, email):
         if email == '':
-            return False, 'invalid_email'
+            return False, None, 'invalid_email'
         dao = StaffDao()
         staff = dao.getStaffByEmail(email)
         if staff == None:
-            return False, 'no_staff'
+            return False, None, 'no_staff'
         else:
-            return staff, 'staff_exists'
+            return True, staff, 'staff_exists'
 
-    def updateStaffAsAdmin(self, eid, admin):
+    def updateStaffAsAdmin(self, eid, form):
+        admin = form['admin']
         if eid == '' or admin == '':
             return False, 'invalid_form'
         if self.staffExists(eid):
             dao = StaffDao()
-            dao.updateStaffAsAdmin(eid, admin)
+            dao.updateStaffAsAdmin(eid, v().stringToBool(admin))
             return True, 'updated_admin'
         else:
             return False, 'no_staff'
 
-    def updateStaffStore(self, eid, store):
+    def updateStaffStore(self, eid, form):
+        store = form['store']
         if eid == '' or store == '':
             return False, 'invalid_form'
         if self.staffExists(eid):
@@ -83,8 +96,8 @@ class StaffHandler:
     def insertStaff(self, form):
         fname = form['first_name']
         lname = form['last_name']
-        eid = form['eid']
-        email = form['email']
+        eid = v().toLower(form['eid'])
+        email = v().toLower(form['email'])
         store = form['store']
         admin = v().stringToBool(form['admin'])
         password = self.generatePassword()
@@ -96,7 +109,10 @@ class StaffHandler:
             StaffDao().insertStaff(fname, lname, eid, admin, email, password, store)
             return True, 'staff_created'
 
-    def deleteStaff(self, eid):
+    def deleteStaff(self, form):
+        eid = v().toLower(form['eid'])
+        if eid == '':
+            return False, 'invalid_form'
         if self.staffExists(eid):
             StaffDao().deleteStaffbyEid(eid)
             return True, 'staff_deleted'
